Fixing the computation of num_mb

+++ DirectProgramming/DPC++FPGA/Tutorials/Features/read_only_cache/src/read_only_cache.cpp
@@ -152,7 +152,8 @@ int main() {
     std::cout.setf(std::ios::fixed);
 
     // Input size in MB
-    constexpr double num_mb = (kNumOutputs * sizeof(uint32_t)) / (1024 * 1024);
+    constexpr double num_mb =
+        (static_cast<double>(kNumOutputs * sizeof(uint32_t))) / (1024 * 1024);
 
     // Report kernel execution time and throughput
     std::cout << "Kernel execution time: " << time_kernel << " seconds\n";
